implement CRUD for board

+++ backend/TasksBoard.Backend/Features/Tasks/Delete.cs
@@ -7,6 +7,9 @@ using FluentValidation;
 
 using MediatR;
 
+using Microsoft.EntityFrameworkCore;
+
+using TasksBoard.Backend.Infrastructure;
 using TasksBoard.Backend.Infrastructure.Context;
 using TasksBoard.Backend.Infrastructure.Errors;
 
@@ -34,19 +37,26 @@ namespace TasksBoard.Backend.Features.Tasks
 
         public class QueryHandler : IRequestHandler<Command>
         {
+            private readonly ICurrentUserAccessor _currentUserAccessor;
             private readonly TasksBoardContext _context;
 
-            public QueryHandler(IDbContextInjector dbContextInjector)
+            public QueryHandler(IDbContextInjector dbContextInjector, ICurrentUserAccessor currentUserAccessor)
             {
+                _currentUserAccessor = currentUserAccessor;
                 _context = dbContextInjector.WriteContext;
             }
 
             public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
             {
-                var task = await _context.Tasks.FindAsync(request.TaskId);
+                var task = await _context.Tasks
+                    .Include(t => t.Owner)
+                    .SingleOrDefaultAsync(t => t.Id == request.TaskId, cancellationToken);
 
                 if (task == null)
                     throw new RestException(HttpStatusCode.NotFound, new { Task = Constants.NOT_FOUND });
+
+                if (!task.Owner.Email.Equals(_currentUserAccessor.GetCurrentUserEmail()))
+                    throw new RestException(HttpStatusCode.BadRequest, new { User = Constants.NO_OWNER });
 
                 _context.Tasks.Remove(task);
                 await _context.SaveChangesAsync(cancellationToken);
